Migration off next theme
Next theme is giving off errors and even trought those errors dont directly impact the website functinality it was desided to still get rid off them by migrating to a custom theme hendler.

Currently Im using cookies-next until a proper support for cookies will be added, until than the dev branch will not be pushed to main due to the fact that there is error with theme saving due to the fact that cookies-next do not support app dir.